+++ Development/ros2/src/qcar2_autonomy/launch/autonomy_planner_launch.py
@@ -28,13 +28,6 @@ def generate_launch_description():
     # }]
 )
     
-    yolo_detector = Node(
-        package='qcar2_autonomy',
-        executable='yolo_detector',
-        name='yolo_detector',
-
-    )
-
     lane_detection = Node(
         package='qcar2_autonomy',
         executable='lane_detection',
@@ -65,7 +58,6 @@ def generate_launch_description():
         trip_planner,
         # lane_detection,
         # lane_stanley_node,
-        yolo_detector,
         #sidewalk_detection,
         ]
     )